# core-common-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "core-common-customer-mgmt",
    "description": "Customer Management Core Application",
    "contact": {
      "name": "Firefly Software Foundation",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/customers",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/parties/{partyId}/party-statuses": {
      "put": {
        "tags": [
          "Party Statuses"
        ],
        "summary": "Update party status",
        "description": "Update an existing party status associated with a party",
        "operationId": "updatePartyStatus",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid party status data"
          },
          "404": {
            "description": "Party status or party not found"
          },
          "200": {
            "description": "Party status successfully updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PartyStatusDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Party Statuses"
        ],
        "summary": "Create party status for a party",
        "description": "Create a new party status associated with a specific party",
        "operationId": "createPartyStatus",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Party status successfully created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PartyStatusDTO"
                }
              }
            }
          },
          "404": {
            "description": "Party not found"
          },
          "400": {
            "description": "Invalid party status data"
          }
        }
      }
    },
...
    "/api/v1/parties/{partyId}/party-statuses/filter": {
      "post": {
        "tags": [
          "Party Statuses"
        ],
        "summary": "Filter party statuses for a party",
        "description": "Retrieve a paginated list of party statuses associated with a specific party based on filtering criteria",
        "operationId": "filterPartyStatuses",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestPartyStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Party not found"
          },
          "200": {
            "description": "Successfully retrieved filtered party statuses",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter request"
          }
        }
      }
    },
...
    "/api/v1/parties/{partyId}/party-statuses/{partyStatusId}": {
      "get": {
        "tags": [
          "Party Statuses"
        ],
        "summary": "Get party status by ID",
        "description": "Retrieve a specific party status associated with a party by its unique identifier",
        "operationId": "getPartyStatusById",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "partyStatusId",
            "in": "path",
            "description": "Unique identifier of the party status",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Party status found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PartyStatusDTO"
                }
              }
            }
          },
          "404": {
            "description": "Party status or party not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Party Statuses"
        ],
        "summary": "Delete party status",
        "description": "Delete a party status associated with a party from the system",
        "operationId": "deletePartyStatus",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "partyStatusId",
            "in": "path",
            "description": "Unique identifier of the party status",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Party status successfully deleted"
          },
          "404": {
            "description": "Party status or party not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PartyStatusDTO": {
        "required": [
          "partyId",
          "statusCode"
        ],
        "type": "object",
        "properties": {
          "partyStatusId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "partyId": {
            "type": "string",
            "format": "uuid"
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED",
              "PENDING",
              "CLOSED"
            ]
          },
          "statusReason": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time"
          },
          "validTo": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "caseInsensitiveStrings": {
            "type": "boolean"
          },
          "includeInheritedFields": {
            "type": "boolean"
          }
        },
        "description": "Filter options for customizing filter behavior"
      },
...
      "PaginationRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "The zero-based page number to retrieve.",
            "format": "int32",
            "example": 0,
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "description": "The number of items per page.",
            "format": "int32",
            "example": 10,
            "default": 10
          },
          "sortBy": {
            "type": "string",
            "description": "The field to sort the results by.",
            "example": "name"
          },
          "sortDirection": {
            "type": "string",
            "description": "The direction of sorting, either ASC or DESC.",
            "example": "DESC",
            "default": "DESC"
          }
        },
        "description": "Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction."
      },
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
      "RangeObject": {
        "type": "object",
        "properties": {
          "from": {
            "type": "object",
            "description": "Start value of the range"
          },
          "to": {
            "type": "object",
            "description": "End value of the range"
          }
        },
        "description": "Represents a range with from and to values"
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "type": "object",
              "description": "The list of items for the current page."
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      },
      "FilterRequestPartyStatusDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PartyStatusDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Filter criteria for party statuses"
      },
...
    }
  }
}